API/src/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Flask-RESTful==0.3.10
google-auth==2.43.0
google-genai==1.52.0
numpy==2.3.5
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
yfinance==0.2.66
